Move VDO statistics code to a separate file
We'll be using this code from the LVM plugin too.
